feat: move integration tests into separate package and add unit tests
