+++ bindings/go/componentversionrepository/integration/go.mod
@@ -1,0 +1,3 @@
+module ocm.software/open-component-model/bindings/go/componentversionrepository/integration
+
+go 1.24.2
