+++ bindings/go/componentversionrepository/integration/testdata/fallback-transport-archive-constructor.yaml
