+++ bindings/go/componentversionrepository/integration/testdata/fallback-transport-archive/artifact-index.json
