+++ bindings/go/componentversionrepository/integration/testdata/transport-archive-constructor.yaml
