+++ bindings/go/componentversionrepository/integration/testdata/transport-archive-copy/artifact-index.json
