+++ bindings/go/componentversionrepository/integration/testdata/transport-archive/artifact-index.json
